Replace removed NARROW_COLUMNS layout mode with TEXT_AUTOSIZING
For reasons which I completely fail to understand, Google elected not to
implement the very useful NARROW_COLUMNS layout mode (double-tap to
reflow text) in Chromium WebView, and instead chose to implement
TEXT_AUTOSIZING (buggy heuristic-based font inflation).  But complaining
isn't going to get it back, and font inflation is better than nothing on
small screens, so replace NARROW_COLUMNS with TEXT_AUTOSIZING to restore
some support for resizing text to be legible on screen.

Change-Id: Ieeb03d38e8a7386e8e8f811c570aa17bfabf4337